requirements/base.txt
Django==5.1.2
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3

psycopg2-binary>=2.9

drf-spectacular>=0.27
drf-spectacular-sidecar>=2024.1